Cleanup warnings

+++ src/include/reactor_epoll.c
@@ -234,7 +234,7 @@ bool reactor_wakeup_pending(Reactor *reactor, GRContext *context) {
   bool notifed = false;
   size_t shard_count = context->shard_set->size;
   ShardSet *shard_set = context->shard_set;
-  for (int i = 0; i < shard_count; i++) {
+  for (size_t i = 0; i < shard_count; i++) {
     if (BITSET64_GET(reactor->soft_notify, i)) {
       Shard *shard = &shard_set->shards[i];
       Reactor *r = shard->reactor;
